requirements.txt
boto3==1.1.4
botocore==1.2.7
datadiff==1.1.6
docutils==0.12
Flask==0.10.1
futures==2.2.0
gunicorn==19.3.0
itsdangerous==0.24
Jinja2==2.8
jmespath==0.9.0
MarkupSafe==0.23
pymongo==3.0.3
python-dateutil==2.4.2
six==1.9.0
Werkzeug==0.10.4
